Move stuff into the symmetriccipher and hash modules
