+++ src/hash/digest.rs
