+++ src/hash/mod.rs
@@ -1,0 +1,13 @@
+pub mod bcrypt;
+pub mod bcrypt_pbkdf;
+pub mod blake2b;
+pub mod digest;
+pub mod hkdf;
+pub mod md5;
+pub mod pbkdf2;
+pub mod rc4;
+pub mod ripemd160;
+pub mod scrypt;
+pub mod sha1;
+pub mod sha2;
+pub mod whirlpool;
