+++ src/hash/scrypt.rs
@@ -23,9 +23,9 @@ use serialize::base64;
 use serialize::base64::{FromBase64, ToBase64};
 
 use cryptoutil::{read_u32_le, read_u32v_le, write_u32_le};
+use hash::pbkdf2::pbkdf2;
+use hash::sha2::Sha256;
 use hmac::Hmac;
-use pbkdf2::pbkdf2;
-use sha2::Sha256;
 use util::fixed_time_eq;
 
 // The salsa20/8 core function.
@@ -410,7 +410,7 @@ pub fn scrypt_check(password: &str, hashed_value: &str) -> Result<bool, &'static
 mod test {
     use std::iter::repeat;
 
-    use scrypt::{scrypt, scrypt_simple, scrypt_check, ScryptParams};
+    use hash::scrypt::{scrypt, scrypt_simple, scrypt_check, ScryptParams};
 
     struct Test {
         password: &'static str,
