+++ src/symmetriccipher/chacha20poly1305.rs
@@ -4,9 +4,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use aead::{AeadEncryptor,AeadDecryptor};
+use symmetriccipher::aead::{AeadEncryptor,AeadDecryptor};
 
-use chacha20::ChaCha20;
+use symmetriccipher::chacha20::ChaCha20;
 use symmetriccipher::SynchronousStreamCipher;
 use poly1305::Poly1305;
 use mac::Mac;
@@ -89,8 +89,8 @@ impl AeadDecryptor for ChaCha20Poly1305 {
 mod test {
   use std::iter::repeat;
 
-  use chacha20poly1305::ChaCha20Poly1305;
-  use aead::{AeadEncryptor,AeadDecryptor};
+  use symmetriccipher::chacha20poly1305::ChaCha20Poly1305;
+  use symmetriccipher::aead::{AeadEncryptor,AeadDecryptor};
   struct TestVector {
     key:   [u8; 32],
     nonce: [u8; 8],
@@ -738,8 +738,8 @@ mod test {
 #[cfg(all(test, feature = "with-bench"))]
 mod bench {
     use test::Bencher;
-    use chacha20poly1305::ChaCha20Poly1305;
-    use aead::{AeadEncryptor, AeadDecryptor};
+    use symmetriccipher::chacha20poly1305::ChaCha20Poly1305;
+    use symmetriccipher::aead::{AeadEncryptor, AeadDecryptor};
 
     #[bench]
     pub fn chacha20poly1305_10(bh: & mut Bencher) {
