+++ src/symmetriccipher/hc128.rs
@@ -180,8 +180,8 @@ impl Decryptor for Hc128 {
 
 #[cfg(test)]
 mod test {
-    use hc128::Hc128;
     use symmetriccipher::SynchronousStreamCipher;
+    use symmetriccipher::hc128::Hc128;
     use serialize::hex::{FromHex};
 
     // Vectors from http://www.ecrypt.eu.org/stream/svn/viewcvs.cgi/ecrypt/trunk/submissions/hc-256/hc-128/verified.test-vectors?rev=210&view=markup
@@ -260,7 +260,7 @@ mod test {
 mod bench {
     use test::Bencher;
     use symmetriccipher::SynchronousStreamCipher;
-    use hc128::Hc128;
+    use symmetriccipher::hc128::Hc128;
 
     #[bench]
     pub fn hc128_10(bh: & mut Bencher) {
